requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
PyMySQL==1.0.2
python-dotenv==0.10.3